chore: refactor big code to smaller functions
